wip setting up to use the orchestrator
made the client app use simplep2p test. right now it's using
codes config to make sure that it's working. now i'll continue
implementing the yaml and dot configs, and eventually have the
orchestrator do all of the lp setup.
when done, we should have the same output as the simplep2p test