added in ability to create a custom lp type
We want to be able to make it so the user doesn't necessarily need to
write code in order to create a simulation. So the client and the
orchestrator handles that, but we want to retain the ability to create
your own non-model-net LPs outside of codes and have more control over
simulation configuration.